fix videoplayer stickersheet
